make sure to execute first camera update
The camera plugin impl executes camera jumps on a tick of first animator. The animator is registered _also_ when it first ticks, but it was registered last. Because the logic to execute the updates checks for the registered animators, and the animators was registered last in that logical block, the first tick of the animation was effectively ignored.