[gestures] mitigate fast fling in downward direction when map is pitched.
By retroducing the pitch limiting factor, we workaround the issue of transforming the map too much in a downward direction when the map is fully pitched. This improves the UX to a more expected behavior. It also improves the UX when using terrain, you will see less tiles missing when flinging.