Integrate the new camera changed coalesced into platform code (#3562)
PR #3525 introduced a new event: camera changed coalesced.

We do not currently expose the new event fully. It is annotated with
`@RestrictTo(Scope.LIBRARY_GROUP_PREFIX)` and also `MapboxExperimental`.
We might expose it later on in 11.13.0-beta.1.

This PR substitutes the previous use of `cameraChanged` event in
scalebar and `MapCamera` plugins.

Previously, with `cameraChanged` event, we were running unnecessary code
in those cases where the camera state changed internally but it was not
the one being rendered in the map. That was clearly visible when doing
an animation (for example `flyTo`) where each animator calls
`setCamera`. Every `setCamera` results in `cameraChanged` event. The new
`cameraChangedCoalesced` will only be triggered once, once all the
`setCamera` are done and the new camera options to be rendered is known.

https://mapbox.atlassian.net/browse/MAPSAND-2116

cc @mapbox/sdk-ci

GitOrigin-RevId: 394adbb4e8c7d40c37ab2326ed1538042955508c