Expose minZoom / maxZoom for annotation managers (#3747)
1. The `"minzoom"` and `"maxzoom"` are the original names in gl, we
needed to support custom kotlin properties names. It has been moved it
to the `property.key` from annotation code generation. We might consider
migrating other generators when needed or change the name.
2. The `isNumberTypeLong` function checked from zoom to be the `Long`
type. There is no need for that as this property supports floating
numbers. The current `Long` types in generated code for `Source`are
preserved to not to introduce breaking changes. However properties in
`AnnotationManager` are `Double`
3. `layer.properties` are hardcoded to keep the changes small and not to
affect other projects. Format for these properties is different as they
are not in `layer.type`

cc @mapbox/sdk-ci

cc @mapbox/maps-android
cc @mapbox/maps-ios

GitOrigin-RevId: 67747155f1d901bad58ba52c2f8f206812000b24