Bump compile and target sdk 35 (#6184)
https://mapbox.atlassian.net/browse/MAPSAND-2272

Decided to do windowOptOutEdgeToEdgeEnforcement to not to overcomplicate
examples.

There is an issue otherwise when content is overlapped by default
supportActionBar. Options would be to
1. handle viewTree to get action bar's height, somehow combine with
window insets listener, update padding/margin.
2. Update each layout for every activity to use custom Toolbar.
3. Use methods from support library, it would require updating them, so
as updating kotlin, so as updating ksp, scripts and the waterflow goes
on.

SDK 35 is the current requirement so SDK 36 is not yet enforced and
there are some changes that would be considered as a separate task.

cc @mapbox/maps-android
cc @mapbox/core-sdk

cc @mapbox/sdk-ci

---------

Co-authored-by: Changelog autocreator bot <actions@github.com>
GitOrigin-RevId: e4bd8fb7405e5d7c594349c742c7db165ed0f0e2